backend/autonhome/requirements.txt
django
djangorestframework==3.14.0
Pyrebase==3.0.27
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
schedule
python-dotenv==1.0.0
django-cors-headers==4.0.0
dj-rest-auth==4.0.0
